Getting rid of developing dir (just use branches), and starting to add in longleaf post-processing
