+++ post_proc/post_proc.sh
@@ -1,0 +1,8 @@
+#!/bin/sh
+
+for filename in $( ls *.gsd )
+do
+
+    sh analyze.sh $filename
+
+done
